使用double作为pos的搞定了，原因是给出的卫星坐标点最后一个和第一个重复，而double由于计算机硬件存储原因可能导致出现.999999类似的情况从而导致叉乘法在最后一个边，也就是两个相同的点（头尾）连起来的时候计算出现奇怪的结果导致错误判定！！！！

+++ Algo_SatelliteCovCalc&MissionPlanning/AllKindOfStruct.h
@@ -10,10 +10,10 @@
 #include "Header.h"
 
 struct EarthPos{
-    float _x = -1;
-    float _y = -1;
+    double _x = -1;
+    double _y = -1;
     EarthPos() = default;
-    EarthPos(float x, float y):_x(x), _y(y){}
+    EarthPos(double x, double y):_x(x), _y(y){}
 };
 //struct Rectangle {
 //    EarthPos left_up;
@@ -82,7 +82,7 @@ struct two_vec//两个向量
     EarthPos end_vec; //裁剪窗口边界向量终点
     //result函数返回叉乘方向
     int result() {
-        float t = start_vec._x*end_vec._y - end_vec._x*start_vec._y;
+        double t = start_vec._x*end_vec._y - end_vec._x*start_vec._y;
         if (t > 0)
             return 1;
         else if (t < 0)
